requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu==2.15.0
pillow
python-multipart
numpy==1.26.*
